requirements.txt
alembic==1.11.2
asyncpg==0.28.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
et-xmlfile==1.1.0
Faker==19.3.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
idna==3.4
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.5
openpyxl==3.1.2
proto-plus==1.22.3
protobuf==4.24.0
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML @ file:///private/tmp/pyyaml-20230717-6270-1iiwi1j/PyYAML-6.0.1
random-password-generator==2.2.0
requests==2.31.0
rsa==4.9
russian-names==0.1.2
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
